+++ bin/update-deps.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 # Update python and npm dependencies
+source .venv/bin/activate
 set -euo pipefail
 poetry update
 poetry show --outdated